Replace ParseWithFactory with a new Flatten parser
Breaking change: It removes the ParseWithFactory
base parser. It could either have a factory
	Dict<String, String> -> String, or
	String -> String.

But, in a parser that inherits from ParseWithFactory,
it was up to the implementer to use the HasFactory
flag before using one of the two factory method.

As alternative, it introduces a Flatten parser. It
received a "flattening" function and an inner parser.
It collapses the output of a parser into a single string.

Since, the most common use case is to parse a line
with a regex and extract the matching group, the
ParseFromRegex parser still uses a factory as a paramater.